requirements.txt
pandas>=1.3.0
numpy>=1.21.0
requests>=2.26.0
sentence-transformers>=2.2.0
faiss-cpu>=1.7.2
